+++ CHANGELOG.md
@@ -1,13 +1,18 @@
 # Change Log
 
+## 1.1.0
+
+* Added the feature to select a macro file and then run it.
+* Some code improvement.
+
 ## 1.0.5
 
-* Add a new `debug` command.
+* Added the feature to launch the visual studio extension host.
 * Add an extension icon.
 
 ## 1.0.4
 
-* Add a new 'openMacroDirectory' command.
+* Added the feature to open macro directories in VSCode.
 * Tested the extension in the VSCode 1.31.1 and lowered the minimum version to 1.31.1.
 
 ## 1.0.3